Musze nauczyc sie uzywania branchy

+++ app/src/main/java/com/example/psalata/moneysaver/incomes/Income.java
@@ -3,14 +3,16 @@ package com.example.psalata.moneysaver.incomes;
 
 import com.example.psalata.moneysaver.R;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Paweł on 2016-02-04.
  */
 public class Income {
-    private Double amount;
+    private BigDecimal amount;
     private String date;
 
-    public Income(Double amount, String date) {
+    public Income(BigDecimal amount, String date) {
         this.amount = amount;
         this.date = date;
     }
